Simplify the right-hand side of the prompt.
The prompt no longer unconditionally shows the shell level and error
code. Now, if in a subshell, a "<" arrow is shown on the left-hand side
(indicating that there's something to go back to, and being just a
subtle and non-intrusive indicator).

If there's an error, we show a red "< $?" on the far-right. If the error
code is 0, we don't show it.