[zsh] Fix the display of arrows in the prompt.
My zsh prompt has had powerline-like arrows for a while, but these were
using an older character that caused massive problems on El Capitan, due
to some issue with a particular font.

I've switched to more appropriate arrows, and added a variable for each,
to make it easier to manage the prompts and change them in the future.