[zsh] Always assume the base shell level is 1 for the prompt.
I had a custom check for iTerm2, with a base shell level of 2. The idea
being that we indicate in the prompt if we're in a subshell. iTerm2
seems to use a base shell level of 1 now, which is either due to a
reboot or to the newer versions of iTerm2. Not sure, but taking this out
for now.