[git] Add new aliases for creating WIP branches.
These make it easy to create a starting point for a WIP feature/fix
branch, where the summary begins with "[WIP]" and the contents contain a
TODO. These commits are empty, and can be easily amended.

`git wip` will create the commit on an existing branch, whereas
`git nbw` will create a new branch with a WIP commit.