[vim] Split .vimrc up into smaller purpose-driven files.
This is an attempt at organizing the .vimrc mess, moving much of the
configuration into separate files.

Plugin configuration now lives in files in config/plugins/. Some macros
are in config/macros/. Other stuff live directly in config/.

This is definitely not perfect, nor a final version of this. I don't
like the macros/ stuff much, and am considering consolidating much of
it. Still, it's hopefully a step in the right direction.