[zsh] Improve management of development projects.
This introduces a `proj` command that can switch to a development
project. This will load a project configuration in `~/.zsh/projects/`
and use it to change the Python venv buildroot, the current directory,
and the iTerm tab color/title and pane badge (all optional).

Tab completion is available for this, and for the existing `buildroot`
command as well.