[zsh] Add tab completion for pydo.
`pydo` is a utility command for running `python`, `pip`, or a
script/tool with a specific set of Python versions. It's useful when
working with environments and projects involving many versions of
Python.

`pydo` itself is currently local to my `bin` directory, but will be made
available as part of the `virtualenv-multiver` package later.